feat: changes blm-preview into own domain, moves wait-function into utils folder
