+++ webshot/src/domain/utils/wait-function.ts
